The animations have all been successfully fixed!  YESSSSSSSSSSSS AHH AAHHAHAHHAHAHAHAHHAHAHAHAHAH.
Very happy.  Anyway, all that had to be done from the last commit was change all of the animations all back to Write Defaults (which they were at by default), and check as well as uncheck "Apply Root Motion" (I don't think that this did anything, I am just putting this in here in case I encounter an issue similar to this and this comment becomes applicable to that scenario).  One thing that also happened that was relatively important was the fact that projectiles now only attack objects that have a trigger collider and are on the "Fighting" layer.  Pretty important, it turns out that the player was being attacked through the DropHandler trigger.